requirements.txt
# Underwater Marine Waste Detection - Dependencies
# Install: pip install -r requirements.txt

# Core
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8

# Configuration & Utilities
PyYAML>=6.0
tqdm>=4.65.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Optional: For Colab compatibility
# ipywidgets>=7.6.0
# google-colab  # Only if running on Colab
